% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_size.R
\name{dir_size}
\alias{dir_size}
\title{Size of directories}
\usage{
dir_size(folder, Recurse = FALSE)
}
\arguments{
\item{folder}{A directory}

\item{Recurse}{Whether to recurse through \code{folder} into subdirectories.}
}
\value{
Size of folder and subfolders
}
\description{
Size of directories
}
