% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutilsInteractive-package.R
\docType{package}
\name{hutilsInteractive-package}
\alias{hutilsInteractive}
\alias{hutilsInteractive-package}
\title{hutilsInteractive: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
